+++ js/data.js
@@ -1,0 +1,17 @@
+import { getRandInt } from "./utils";
+
+function getPhotos() {
+  const photos = [];
+  for (let i=0; i<25; i++) {
+    photos.push({
+      id: i+1,
+      url: `photos/${i+1}.jpg`,
+      description: `my fav photo #${i+1}`,
+      likes: getRandInt(15, 200),
+      comments: getRandInt(0, 200)
+    })
+  }
+  return photos;
+}
+
+export {getPhotos};
