+++ js/utils.js
@@ -1,0 +1,9 @@
+function getRandInt(from, to) {
+  return Math.floor(Math.min(from) + Math.random() * Math.abs(to-from));
+}
+
+function isStringFits(string, maxLength) {
+  return string.length <= maxLength;
+}
+
+export {getRandInt, isStringFits};
